requirements.txt
black
flake8
mypy
pytest
pillow